% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregate.holdOutPredictions}
\alias{aggregate.holdOutPredictions}
\title{Aggregate hold-out predictions}
\usage{
\method{aggregate}{holdOutPredictions}(x, ...)
}
\arguments{
\item{x}{an object of class \code{holdOutPredictions} returned from \code{\link{holdOutPredictions}}.}

\item{...}{currently unused.}
}
\description{
Aggregates hold-out predictions which have been
extracted from a \code{\link{trainOcc}} object with the function \code{\link{holdOutPredictions}}.
}
\examples{
\dontrun{
toy <- threeGaussians()
index <- createResample(toy$tr[, 1], times=2)
oc <- trainOcc(x=toy$tr[, -1], y=toy$tr[, 1],
               index=index,
               tuneGrid=expand.grid(sigma=c(0.1,1), ### not so large grid
                                    cNeg=2^seq(-5, -5, 5),
                                    cMultiplier=2^seq(4, 12, 4)))
hop <- holdOutPredictions(oc)
boxplot(list(pos.r1=hop$pos$Resample1, pos.r2=hop$pos$Resample2,
             un.r1=hop$un$Resample1, un.r2=hop$un$Resample2))
hop.agg <- aggregate(hop)
boxplot(list(pos=hop.agg$pos, un=hop.agg$un))
}
}

