% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregate}
\alias{aggregate}
\title{Aggregate held out predictions... .}
\usage{
aggregate(x, ...)
}
\arguments{
\item{x}{a list as returned from \code{\link{heldOutPredictions}}.}

\item{...}{currently unused.}
}
\description{
...
}
\examples{
### to do
}

