% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bananas_trainset.R
\name{bananas_trainset}
\alias{bananas_trainset}
\title{Title}
\usage{
bananas_trainset(nP = 50, nU = 100, nN = 0, seed = NULL)
}
\arguments{
\item{nP}{}

\item{nU}{}

\item{nN}{}

\item{seed}{}
}
\description{
Title
}

