% Generated by roxygen2 (4.0.1): do not edit by hand
\name{confusionMatrix4latex}
\alias{confusionMatrix4latex}
\title{confusionMatrix4latex}
\usage{
confusionMatrix4latex(obj, obj2 = NULL, obj3 = NULL, obj4 = NULL, ...)
}
\arguments{
\item{obj}{an object of class "confusionMatrix" \code{\link{confusionMatrix}}.}

\item{obj2}{another object like obj.}

\item{obj3}{another object like obj.}

\item{obj4}{another object like obj.}

\item{...}{arguments that can be passed to \code{\link{xtable}}, e.g. cap and .}
}
\value{
one or more confusion matrices in latex table format.
}
\description{
Given an object of class confusionMatrix (from caret), a slim xtable is printed to be used in latex documents.
}
\details{
...
}
\examples{
### to do
}
\author{
Benjamin Mack
}
\seealso{
\code{\link{confusionMatrix}}, \code{\link{xtable}}
}

