% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataForSummaryFunction}
\alias{dataForSummaryFunction}
\title{dataForSummaryFunction}
\usage{
dataForSummaryFunction(x, threshold = 0)
}
\arguments{
\item{x}{an object of class \code{\link{holdOutPredictions}}}

\item{threshold}{threshold used to binarize the data}
}
\value{
a data frame that can be passed to the \code{data} argument of \code{\link{puSummary}} function.
}
\description{
Creates a data frame from an object of class \code{holdOutPredictions} that can be passsed to the \code{data} argument of \code{\link{puSummary}} function.
}
\examples{
\dontrun{
data(bananas)
model <- trainOcc(x=bananas$tr[, -1], y=bananas$tr[, 1], method="ocsvm")
hop <- holdOutPredictions(model) # returns aggregated
d <- dataForSummaryFunction(hop)
hop <- holdOutPredictions(model, partition = 1)
d <- dataForSummaryFunction(hop)
}
}

