% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluateAtTh}
\alias{evaluateAtTh}
\title{Model evaluation / accuracy assessment at one particular threshold.}
\usage{
evaluateAtTh(x, th, ...)
}
\arguments{
\item{x}{an object of class \code{ModelEvaluation},}

\item{th}{the threshold}

\item{...}{arguments passed to \code{\link[dismo]{evaluate}} from the \code{dismo} package.}
}
\value{
an object of class ModelEvaluationAtTh which returns formated confusion matrix and accuracy statistics when printed.
}
\description{
Calculation of accuracy for specified thresholds. This returns an
object which, when printed, returns the confusion matrix and some model evaluation
statistics.
}
\examples{
# get an ModelEvaluation object
ev <- dismo::evaluate( p=rnorm(25, mean = 1), a=rnorm(100, mean = -1) )
# use evaluateAtTh in order to get an
evAt0 <- evaluateAtTh(ev, th=0)
evAtMaxK <- evaluateAtTh(ev, th='max.kappa')
# print the confusion matrices and some model evaluation statistics
evAt0
evAtMaxK
}

