% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_trainOcc.R
\name{fake_trainOcc_bsvm}
\alias{fake_trainOcc_bsvm}
\title{Title}
\usage{
fake_trainOcc_bsvm(results, metric = NULL, control_method = "none")
}
\arguments{
\item{results}{}

\item{metric}{}

\item{control_method}{}
}
\description{
Title
}

