% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featurespace.R
\name{featurespace}
\alias{featurespace}
\title{Plot method for \code{trainOcc} object}
\usage{
featurespace(x, thresholds = NULL, x.tr = NULL, positive = NULL,
  borders = NULL, nCells = c(100, 100), main = NULL, column = NULL,
  expandColors = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{trainOcc}}.}

\item{thresholds}{a scalar or numeric vector used to plot additional contour lines.}

\item{x.tr}{data to be plotted. must have a column y with labels.}

\item{positive}{the class of interest in \code{x.tr$y}}

\item{borders}{names list (x, y) with two vectors of length two specifiying the
borders of the feature space in horizontal/x- and vertical/y- directions.}

\item{nCells}{number of cells in horizontal/x- and vertical/y- directions.}

\item{main}{character string for the plot title}

\item{column}{select a column in case the model \code{x} is not a \code{\link{trainOcc}} object and
returns more then one column, default is \code{1}.}

\item{expandColors}{if the predictive values are not positive and negative numbers (as in case of the ocsvm and biasedsvm) set this argument to \code{FALSE}}

\item{...}{other arguments that can be passed to \code{\link{predict}}.}
}
\value{
a plot of the feature space
}
\description{
Plot the decision values of a \code{\link{trainOcc}} object in its two dimensional feature space.
}
\details{
featurespace

This function is an adaptation of \code{plot.ksvm} from the \code{kernlab} package.
}
\examples{
\dontrun{
data(bananas)
tr <- bananas$tr
### underfitted model
oc <- trainOcc ( x = tr[, -1], y = tr[, 1],
                 tuneGrid=expand.grid(sigma=0.1,
                                      cNeg=.1,
                                      cMultiplier=100))
featurespace(oc, th=0)

### overfitted model
oc <- trainOcc ( x = tr[, -1], y = tr[, 1],
                 tuneGrid=expand.grid(sigma=10,
                                      cNeg=.1,
                                      cMultiplier=100))
featurespace(oc, th=0)
}
}

