% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateThresholds.R
\name{generateThresholds}
\alias{generateThresholds}
\title{Generate thresholds between percentiles from unlabeled and positive data}
\usage{
generateThresholds(data, q_rng = c(0.5, 0.8), nTh = 50)
}
\arguments{
\item{data}{A matrix or data frame with column "obs" and "pos".
"obs": factor with "un" and "pos" levels.
"pos": numeric with the likelihood of the positive class.}

\item{nTh}{Number of thresholds to be generated between the minimum and the maximum.}

\item{q_min}{Minimum threshold as quantile of the unlabeled samples in data.}

\item{q_max}{Maximum threshold as quantile of the positive samples in data.}
}
\value{
vector of thresholds
}
\description{
Generate thresholds between percentiles from unlabeled and positive data
}

