% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getModelInfoOneClass.R
\name{getModelInfoOneClass}
\alias{getModelInfoOneClass}
\title{Models available in \code{\link{trainOcc}}}
\usage{
getModelInfoOneClass(model = NULL, regex = TRUE, ...)
}
\arguments{
\item{model}{a character string of a defined one-class classification method.
So far: ocsvm, biasedsvm, maxent.}

\item{regex}{a logical: should a regular expressions be used? If FALSE, a simple match is conducted against the whole name of the model.}

\item{...}{options to pass to grepl}
}
\value{
a list containing one or more lists of the standard model information
}
\description{
getModelInfoOneClass is adopted from \code{\link{getModelInfo}} but returns
models available in the package \code{\link{trainOcc}}.
}
\examples{
getModelInfoOneClass("biasedsvm")
}
\seealso{
\code{\link{getModelInfo}}
}

