% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histUev.R
\name{histUev}
\alias{histUev}
\title{Diagnostic distributions plot for a \code{\link{trainOcc}} object and threshold dependent accuracy.}
\usage{
histUev(model, ev, yLimits = NULL, ...)
}
\arguments{
\item{model}{an object of class \code{\link{trainOcc}}.}

\item{ev}{an object of class \code{\link{ModelEvaluation}}.}

\item{yLimits}{specify the minimum and maximum points in the density scale of to which the accuracy scale should be projected. This is useful if you add the plot to another one, e.g. a histogram plot (\code{\link{hist.trainOcc}}), whose  y-axis scale is not suitable for the accuracy plot (i.e. if it is not between 0 and 1).}

\item{...}{arguments that can be passed to \code{\link{hist.trainOcc}}, e.g. \code{pred}}
}
\value{
yLimits (invisible) and the diagnostic plot with threshold depenent accuracy.
}
\description{
The histogram of the predicted unlabeled data is shown together with the 
hold-out predictions of the positive and unlabeled traning data (boxplots). The threshold dependent accuracy (user's/producer's accuracy and kappa) is superimposed.
}
\details{
histUev
}

