% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_default_fc.R
\name{maxent_default_fc}
\alias{maxent_default_fc}
\title{maxent_default_fc}
\usage{
maxent_default_fc(nPos)
}
\arguments{
\item{nPos}{number of positive training samples}
}
\description{
...
}
\examples{
\dontrun{
# lower and upper bounds
maxent_default_fc(c(-Inf, 9, 10, 14, 15, 79, 80, Inf))
}
}

