% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_is_default.R
\name{maxent_is_default}
\alias{maxent_is_default}
\title{Test if maxent parameters are the default}
\usage{
maxent_is_default(nPos, beta, fc)
}
\arguments{
\item{nPos}{Number of positive samples the default depends on.}

\item{beta}{Beta}

\item{fc}{Feature class}
}
\value{
TRUE of the nPos, beta, fc are maxxent default settings
}
\description{
Test if maxent parameters are the default
}

