% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxentThresholds.R
\name{maxentThresholds}
\alias{maxentThresholds}
\title{Extract logistic output thresholds.}
\usage{
maxentThresholds(x)
}
\arguments{
\item{x}{an object of class \code{MaxEnt} as returned from \code{\link[dismo]{maxent}} or \code{trainOcc} as returned from \code{\link{trainOcc}} fitted with the method \code{maxent}.}
}
\description{
Extract several logistic output thresholds from an object of class \code{MaxEnt} as returned from \code{\link[dismo]{maxent}} or \code{trainOcc} fitted with the method \code{maxent}.
}
\examples{
\dontrun{
### still missing ...
}
}

