% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentThresholds_shortNames.R
\name{maxentThresholds_shortNames}
\alias{maxentThresholds_shortNames}
\title{Map long to short threshold names.}
\usage{
maxentThresholds_shortNames(x = NULL)
}
\arguments{
\item{x}{NULL (default) or a character vector. 
In the first case the short and long names are printed.
In the last case the (matching) long names in x are converted 
in short names.}
}
\description{
...
}
\examples{
\dontrun{
longnames <- c("X1", 
               "Fixed.cumulative.value.10.logistic.threshold", 
               "X2")
maxentThresholds_shortNames(longnames)
}
}

