% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puSummaryThLoop.R
\name{maximize_puMetric}
\alias{maximize_puMetric}
\title{Title}
\usage{
maximize_puMetric(x, metric)
}
\arguments{
\item{x}{Data frame containing pu-metric values (columns) 
for a sequence of thresholds (rows).}

\item{metric}{A character (vector). The metrics to be 
maximized. Must exist as column name in \code{x}.}
}
\description{
Title
}

