% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modelPosition}
\alias{modelPosition}
\title{Get parameter(s), result table row and rank of a model.}
\usage{
modelPosition(x, modParam = NULL, modRow = NULL, modRank = NULL,
  by = NULL, decreasing = TRUE)
}
\arguments{
\item{x}{an object as returned from train.}

\item{modParam}{data frame with the parameters of the model.}

\item{modRow}{the row or index of the model in the
model-selection table.}

\item{modRank}{the rank of the model after sorting by \code{by}.}

\item{by}{a character or character vector specifying by which column of the \code{x$results} table to sort.
If \code{NULL} the performance metric is taken from the \code{train} object.}

\item{decreasing}{only when \code{modRank} is used. \code{TRUE} (default) to sort in decreasing order.
can be a vector if \code{by} is a vector.}
}
\value{
a list with the parameter, row index and rank of the model.
}
\description{
Given the parameter(s), results table row, or rank of a particular metric of the results table,  the corresponding
 two informations are extracted.
}
\examples{
data(bananas)
oc <- trainOcc(x = bananas$tr[, -1], y = bananas$tr[, 1],
               tuneGrid=expand.grid(sigma=c(.1, 1),
                                    cNeg=c(0.1, 1),
                                    cMultiplier=c(10, 100)))
## get the position (in the results table) and parameters of model which is
## on rank 3 according to the performance metric \\'puAuc\\'
mp3 <- modelPosition(oc, modRank=3, by="puAuc")
}

