% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nicergb}
\alias{nicergb}
\title{nicergb}
\usage{
nicergb(img, r = 1, g = 2, b = 3, stretch = 2, readjust = FALSE,
  fixed = NULL, nodata = NULL)
}
\arguments{
\item{img}{RasterBrick or RasterStack}

\item{r}{integer. Index of the Red channel, between 1 and nlayers(x)}

\item{g}{integer. Index of the Green channel, between 1 and nlayers(x)}

\item{b}{integer. Index of the Blue channel, between 1 and nlayers(x)}

\item{stretch}{numeric. linear stretch to be applied. Values between 0 and 50
describe proportion (in percent) to be cut away from lower and upper quantile.}

\item{readjust}{logical. If \code{TRUE}, the channel with the largest range is scaled
to its maximum. The ranges of the other two channels are stretched but
maintaining the relative proportion.}

\item{fixed}{matrix of 3 x 2 with thresholds for RGB}

\item{nodata}{numeric. no data value to be ignored.}
}
\value{
A list with the following objects:
\itemize{
\item{original.range}{original range of greyscale values}
\item{plotted.range}{the plotted range of greyscale values}
\item{color.stretch}{minimum and maximum color values}
}
}
\description{
Pretty Red-Green-Blue plot based on three layers
}
\details{
The \code{nicergb} function is a wrapper for \code{\link{plotRGB}}
in package package \code{raster}.
}
\examples{
\dontrun{
## Load image
img <- brick(system.file("external/rlogo.grd", package="raster"))
## Plot as RGB
nicergb (img)
## Cut the upper and lower 20\% of the histogram and stretch the rest
nicergb (img, stretch = 20)
}
}
\author{
Sebastian Schmidtlein
}
\seealso{
\code{\link{plotRGB}}
}

