<<<<<<< 66340981c258d3f228091eb35bef1f1d8e338642
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmm_selection_criteria.R
=======
% Generated by roxygen2 (4.0.1): do not edit by hand
>>>>>>> nmm_functions
\name{nmm_selection_criteria}
\alias{nmm_selection_criteria}
\title{nmm_selection_criteria}
\usage{
nmm_selection_criteria(x, un = NULL, nmm = NULL, modRow = NULL,
<<<<<<< 66340981c258d3f228091eb35bef1f1d8e338642
  predU = NULL, predFromNumm = FALSE, fnamePlot = NULL)
=======
  fnamePlot = NULL)
>>>>>>> nmm_functions
}
\arguments{
\item{x}{an object of class \code{trainOcc}}

\item{un}{...}

\item{nmm}{the normal mixture (density) models (see also \code{\link[Mclust]{densityMclust}}}

\item{modRow}{the model in \code{x} used to predict the data with which \code{nmm} has been estimated. If \code{null} the \code{finalModel} (\code{x$finalModel}) is used.}

\item{fnamePlot}{if given the diagnostic is saved to the specified filename (defualt is \code{NULL})}
}
\description{
runs a couple of checks which can be used in order to select a model.
}

