% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.ModelSelectionEvaluation}
\alias{plot.ModelSelectionEvaluation}
\title{Plot PU-performance vs. evaluation accuracies to evaluate model selection.}
\usage{
\method{plot}{ModelSelectionEvaluation}(x, fromEvaluation = "max.kappa",
  fromTrain = x$metric, rowNumbers = FALSE, ...)
}
\arguments{
\item{x}{an object of class ModelSelectionEvaluation
(see \code{\link{evaluate}}).}

\item{fromEvaluation}{character specifying what to plot on the y-axis. can be one of the
slots of an evaluation object (see \code{\link{ModelEvaluation-class}}).}

\item{fromTrain}{character specifying what to plot on the x-axis. can be one of the
\code{colnames(x$train)}. default is the metric used when calling \code{oneClass}.}

\item{rowNumbers}{if \code{TRUE} the points are annotated with the row-number (in the model selection table)
of the models,}

\item{...}{other arguments that can be passed to \code{\link{plot}}.}
}
\description{
...
}
\examples{
### to do
}

