% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_ocsvm.R
\name{plot_consistent_ocsvm}
\alias{plot_consistent_ocsvm}
\title{Plot model complexity (sigma) versus FNR of consistency tuned OCSVM}
\usage{
plot_consistent_ocsvm(x)
}
\arguments{
\item{x}{object returned by \code{\link{consistent_ocsvm}}}
}
\description{
Plot model complexity (sigma) versus FNR of consistency tuned OCSVM
}
\examples{
#' data(bananas)
idx.p <- sample(which(bananas$y[]==1), 50)
x.p <- bananas$x[][idx.p, ]
mod <- consistent_ocsvm(x.p, sigma_thr=2)
# plot_consistent_ocsvm(mod)
}

