% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_puPpvVsTpr}
\alias{plot_puPpvVsTpr}
\title{plot_puPpvVsTpr}
\usage{
plot_puPpvVsTpr(x, cexScaling = "absLogScaled", metric = x$metric,
  cex = 1, identifyPoints = FALSE, ...)
}
\arguments{
\item{x}{a oneClass object}

\item{cexScaling}{scale the size of the points, usually according to a
performance metrics}

\item{metric}{which metric should to be used for scaling}

\item{cex}{multiply cex values
plot_TprVsPuPpv}
}
\description{
Plot the true positive rate and the PU-positive predictive value.
}

