% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_thDepNPPvsTPR}
\alias{plot_thDepNPPvsTPR}
\title{Diagnostic distributions plot for a \code{\link{trainOcc}} object.}
\usage{
\method{hist}{trainOcc}(x, predUn = NULL, add = FALSE,
  scaleToRange = NULL)
}
\arguments{
\item{x}{an object of class \code{\link{trainOcc}}.}

\item{predUn}{a vector of unlabeled predictions (if \code{NULL} \code{x$predUn} is used, if existing).}

\item{th}{draw vertical lines in the histogram, indication for a threshold.}

\item{cab}{for a color-coded histogram a list with
elements \code{colors} (vector of \code{R} colors, length n) and \code{breaks} (vector of numeric values, length n+1).}

\item{main}{a title for the plot. if not given the parameters of the model are added.}

\item{ylim}{the y limits of the plot.}

\item{...}{other arguments that can be passed to \code{\link{plot}}.}
}
\value{
Diagnostic distributions plot.
}
\description{
Plots the true positive rate (TPR) and the probability of negative prediction
(1-PPP, probability of positive prediction) over the complete range of possible thresholds.
}
\details{
plot_thDepNPPvsTPR
}
\examples{
data(bananas)
### a good model
oc <- trainOcc (x = bananas$tr[, -1], y = bananas$tr[, 1],
                tuneGrid=expand.grid(sigma=1,
                                     cNeg=0.0625,
                                     cMultiplier=64))
pred <- predict(oc, bananas$x[])
plot_thDepNPPvsTPR(oc, pred)
### or added to the histogram plot
h <- hist(oc, pred, th=0)
plot_thDepNPPvsTPR(oc, pred, add=TRUE, scaleToRange=c(0, h$ylim[2]) )
}

