% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pnpAt.R
\name{pnpAt}
\alias{pnpAt}
\title{pnpAt}
\usage{
pnpAt(x, modRow = NULL, u = NULL)
}
\arguments{
\item{x}{an object of class \code{\link{trainOcc}}}

\item{modRow}{the row in the results table corresponding to the model for which to derive the ppp.}

\item{u}{a data which can be predicted with the \code{x}. If given the predictions made on \code{u} are used for the calculation instead of the held out predictions stored in \code{x}}
}
\value{
the lower whisker threshold and the corresponding probability of positive prediction at the Lower whisker
}
\description{
Calculate the probability of negative prediction at the minimum/lower whisker of the positive hold-out predictions for one or more models (parameter settings) stored in the \code{\link{trainOcc}} object.
}

