% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.oneClass}
\alias{predict.oneClass}
\title{Predict method for \code{\link{oneClass}} object}
\usage{
\method{predict}{oneClass}(object, newdata, type = "prob",
  allowParallel = TRUE, returnRaster = TRUE, mask = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{oneClass}} object}

\item{newdata}{new data to predict on.}

\item{type}{default is 'prob' which however returns the continuous decision values (see note!)}

\item{allowParallel}{...}

\item{returnRaster}{default \code{TRUE}}

\item{mask}{if given and if \code{returnRaster=TRUE} only predictions of the valid cells are returned.}

\item{...}{other arguments that can be passed to \code{\link{predict.train}}.}
}
\value{
committee predictions and if desired predictions of individual classifiers.
}
\description{
... .
}
\note{
\code{type='probs'} does NOT return probabilities but the continuous
decision values of the classifier, e.g. distances in the case of the biased svm) and not probabilities!
}
\examples{
# to do
}

