% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ModelEvaluationAtTh.R
\name{print.ModelEvaluationAtTh}
\alias{print.ModelEvaluationAtTh}
\title{Print a \code{ModelEvaluationAtTh} object.}
\usage{
\method{print}{ModelEvaluationAtTh}(x, digits = 0, ...)
}
\arguments{
\item{x}{an object of class \code{ModelEvaluationAtTh}}

\item{digits}{number of digits to be printed}

\item{...}{other arguments that be passed to \code{\link{print}}.}
}
\description{
Print the confusion matrix, User's and Producer's Accuracies  (UA and PA), Overall Accuracy (OA), the Area under the ROC Curve (AUC) and the Kapp(K).
}

