% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ModelSelectionEvaluation}
\alias{print.ModelSelectionEvaluation}
\title{Print an object of class \code{ModelSelectionEvaluation}}
\usage{
\method{print}{ModelSelectionEvaluation}(x, by = "mxK.K", digits = 2, ...)
}
\arguments{
\item{x}{an object of class \code{ModelSelectionEvaluation}}

\item{by}{by which column should the table be sorted, default: \code{max.kappa}}

\item{digits}{number of digits to be printed}

\item{...}{options passed to the generic print method.}
}
\description{
A table is printed with test and PU- performance metrics for the different evaluated models.
Test metrics: \cr
auc: auc
mxK.K: kappa at the threshold which leads to the maximum kappa
mxK.OA: Overall Accuracy at the threshold which leads to the maximum kappa
...
}

