% Generated by roxygen2 (4.0.1): do not edit by hand
\name{puFactor}
\alias{puFactor}
\title{puFactor}
\usage{
puFactor(y, positive = NULL)
}
\arguments{
\item{y}{a vector of labels}

\item{positive}{the label of the positive class in y. if not given the id with the smallest
frequency is used.}
}
\value{
an ordererd factor with the levels un<pos
}
\description{
Convert a vector in a ordert factor with levels un<pos as required when using
the one-class classiifiers from \code{oneClass}.
}
\examples{
puFactor(rep( c( -1, -2, 1 ), c( 5, 10, 3 ) ), positive=1)
}

