% Generated by roxygen2 (4.0.1): do not edit by hand
\name{puSummary}
\alias{puSummary}
\title{puSummary}
\usage{
puSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{a data frame or matrix with columns \code{obs}, \code{pred}, and \code{pos}.
The first two are the binary observed, predicted outcomes and the latter one the
continous outcome for the positive class.}

\item{lev}{a character vector of factors levels for the response (default is \code{NULL}).}

\item{model}{a character string for the model name (as taken form the method argument of train (default is \code{NULL}).}
}
\value{
A vector of performance estimates.
}
\description{
... .
}
\details{
... .
}
\examples{
\dontrun{
### create a PU-data example of random numbers
P <- rnorm( 30, 2 )
U <- c( rnorm( 80 ), rnorm( 20, 2 ) )
d <- data.frame( obs = puFactor( rep( c( 1, 0 ), c( 30, 100 ) ), positive = 1 ),
                 pred = puFactor( c(P, U)>0, positive = TRUE ),
                 pos = c( P, U ) )
puSummary(d)
}
}
\seealso{
\code{\link{postResample}}
}

