% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resamples}
\alias{resamples}
\alias{resamples.oneClass}
\title{... .}
\usage{
\method{resamples}{oneClass}(x, modParam = NULL, modRow = NULL,
  modRank = NULL, metric = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{oneClass}}.}

\item{modParam}{a data frame with the desired modParameters.}

\item{modRow}{the index of the model, i.e. the row in the  \code{object$results} table.}

\item{modRank}{the model at the modelRank-th position after sorting after the performance metric. The performance metric can be specified by \code{by}.}

\item{metric}{if modRank is used, which metric should be used for the ranking?}

\item{...}{...}
}
\description{
... .
}
\examples{
\dontrun{
data(bananas)
oc <- oneClass(x=bananas$tr[, -1], y=bananas$tr[, 1],
               tuneGrid=expand.grid(sigma=c(0.1,1), ### not so large grid
                                    cNeg=2^seq(-5, 10, 3),
                                    cMultiplier=2^seq(4, 15, 2)))
# visualize resamples of the ten models with highest (mean) puAuc values
resamps <- resamples(oc, modRank=1:10, metric="puAuc")
bwplot (resamps, metric='puF')
}
}

