% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sort.train}
\alias{sort}
\alias{sort.train}
\title{Sort or order the model selection table of a \code{\link{train}} or
\code{\link{oneClass}} object.}
\usage{
\method{sort}{train}(x, decreasing = TRUE, by = NULL, printTable = TRUE,
  rows = NULL, cols = NULL, digits = min(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class \code{\link{train}} or \code{\link{oneClass}}.}

\item{by}{a character or character vector specifying by which columns to sort.
If \code{NULL} the performance metric is taken from the \code{train} object.}

\item{decreasing}{\code{TRUE} (default) to sort in decreasing order. can be a vector if \code{by} is a vector.}

\item{printTable}{\code{TRUE} for printing the  table, else \code{FALSE}}

\item{rows}{index of the rows to be printed (ignored if \code{printTable=FALSE}). default(\code{NULL}) prints all.}

\item{cols}{index of the cols to be printed (ignored if \code{printTable=FALSE}). default(\code{NULL}) prints all.}

\item{digits}{the number of digits to print}

\item{...}{not currently used.}
}
\description{
The model selection table lists the parameters of the models
evaluated during model selection together with the calculated PU performance metrics.
This function is useful to find and compare the high ranked models.
}
\examples{
### to do
}

