% Generated by roxygen2 (4.0.1): do not edit by hand
\name{threeGaussians}
\alias{threeGaussians}
\title{Two-dimensional toy data consisting of four Gaussian
distributions.}
\usage{
threeGaussians(n = 200, d = 2:4, n.tr = c(20, 100), seed = NULL)
}
\arguments{
\item{n}{...}

\item{d}{...}

\item{n.tr}{...}

\item{seed}{...}
}
\value{
a list of the toy data set with two data frames (\code{tr} and
\code{te}) both containing the columns \code{y}, \code{x1}, and \code{x2}.
in the training set the column \code{y} in \{0, 1\} indicates if a sample
belongs to the positive class (1) or if it is unlabeled (0), and in the test
set \code{te} if it belongs to the positive class (1) or the negative class
(0). \code{x1} and \code{x2} are the features or predictors.
}
\description{
Two-dimensional toy data consisting of four Gaussian
distributions.
}
\examples{
\dontrun{
toy <- threeGaussians(seed=9)

### plot the test set
par(mfrow=c(1,2))
plot( toy$te[ , -1 ], pch = c( 4, 16 )[ toy$te$y + 1 ] )
legend('topright', legend=c('pos', 'neg'), pch=c(16, 4))

### plot the training set
plot( toy$tr[ , -1 ], pch = c( 4, 16 )[ toy$tr$y + 1 ] )
legend('topright', legend=c('pos', 'un'), pch=c(16, 4))

### train a biased svm model and plot the trained model in the feature space
oc <- oneClass(toy$tr[,-1], toy$tr[,1])
featurespace(oc)

### train a maxent model and plot the trained model in the feature space
oc <- oneClass(toy$tr[,-1], toy$tr[,1], method='maxent')
}
}

