% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update.trainOcc}
\alias{update}
\alias{update.trainOcc}
\title{Update or Re-fit the}
\usage{
\method{update}{trainOcc}(object, u = NULL, modParam = NULL,
  modRow = NULL, modRank = NULL, by = NULL, newMetric = NULL,
  aggregatePredictions = FALSE, mask = NULL, newMetricsOnly = FALSE, ...)
}
\arguments{
\item{object}{a object of class trainOcc}

\item{u}{the unlabeled samples which should be predicted with the updated model. Can be a matrix, data.frame, \code{raster}*, or \code{rasterTiled} object.}

\item{modParam}{a data frame with the desired modParameters.}

\item{modRow}{the index of the model, i.e. the row in the  \code{object$results} table.}

\item{modRank}{the model at the modelRank-th position after sorting after the performance metric. The performance metric can be specified by \code{by}.}

\item{by}{when the modRank is given, the models are ranked by the metric given here. if \code{NULL} the metric in \code{object$metric} is used.}

\item{newMetric}{a function that meet the requirements of a \code{summaryFunction} (see ...).
This updates the \code{trainOcc$train$results} data frame with the new metric(s) returned from function.
This can only be done when \code{savePredictions} has been set to \code{TRUE}.}

\item{aggregatePredictions}{The efault is \code{FALSE}, which means that the performance metric is calculated for
each set of hold-out predictions and the metrics are then aggregated. The mean and the standard deviation is returned in the
\code{object$results} table. If \code{TRUE} the held out predictions are first aggregated and the \code{newMetric} is calculated once.
Of course, no standard deviation of the performance metric can be calculated.
Metrics calculated this way get the suffix \code{AP} in the \code{object$results} table.}

\item{mask}{a mask can be given if \code{u} is a \code{raster*} object.}

\item{newMetricsOnly}{logical with default set to \code{FALSE}. Set to \code{TRUE} if the metrics already contained in the results table should be removed.}

\item{...}{other arguments that can be passed to update.train}
}
\value{
an updated trainOcc object.
}
\description{
Over-write the tuning modParameter selection process of the
d2oneMethod (\code{\link{trainOcc}}, \code{\link{update}}) and
re-classify u with the new model.
}

