% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_puSV.R
\name{add_puSV}
\alias{add_puSV}
\title{add_puSV}
\usage{
add_puSV(x)
}
\arguments{
\item{x}{object of class \code{\link{trainOcc}}}
}
\value{
the same object as the input but with columns \code{x$results$nSV} and \code{x$results$puSV} added. \code{nSV} is the number of support vectors and and \code{puSV=(tpr*100)/nSV}.
}
\description{
Add the performance criteria \code{puSV} to the results table of an object of class \code{\link{trainOcc}}.
}
\references{
Muñoz-Marí, Jordi and Bovolo, Francesca and Gómez-Chova, Luis and Bruzzone, Lorenzo and Camp-Valls, Gustavo (2010): 
Semisupervised one-class support vector machines for classification of remote sensing data.
IEEE TGaRS, 48, 8.\cr
}

