% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createMultiFoldsPu}
\alias{createMultiFoldsPu}
\title{createMultiFoldsPu}
\usage{
createMultiFoldsPu(y, k, times = 5, positive = NULL, index.indep = NULL,
  seed = NULL)
}
\arguments{
\item{y}{a vector of outcomes for the positive and the negative class}

\item{k}{an integer for the number of folds (applied to the positive class)}

\item{times}{the number of repetitions to create}

\item{positive}{the positive class in y. if empty the label with the smaller
frequency is assumed to be the positive class.}

\item{index.indep}{optional, the elements in y which should always been used
in the test group. If not given and the indices are passed to trainOcc they are
randomly sampled frm the argument \code{u}.}

\item{seed}{an integer in order to set a seed point}
}
\description{
Create test/training partition for PU data.
The positive data is split into k groups while the unlabeled data is used
completely in every fold. If you want to use standard cross-validation use
\code{\link{createFolds}}, which served as template for this function.
}
\seealso{
\code{\link{createMultiFoldsPu}}, \code{\link{createFolds}}
}

