% Generated by roxygen2 (4.0.1): do not edit by hand
\name{featurespace}
\alias{featurespace}
\title{featurespace}
\usage{
featurespace(x, thresholds = NULL, positive = NULL, x.tr = NULL,
  borders = NULL, nCells = c(100, 100), main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{oneClass}}.}

\item{thresholds}{a scalar or numeric vector used to plot additional contour lines.}

\item{positive}{class of interest}

\item{x.tr}{the training data}

\item{borders}{names list (x, y) with two vectors of length two specifiying the
borders of the feature space in horizontal/x- and vertical/y- directions.}

\item{nCells}{number of cells in horizontal/x- and vertical/y- directions.}

\item{...}{other arguments that can be passed to \code{\link{predict}}.}
}
\value{
...
}
\description{
Plot the decision values of a \code{\link{oneClass}} object in its two dimensional feature space.
}
\details{
This function is an adaptation of the \code{\link{plot.ksvm}}.
}

