% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hist.oneClass}
\alias{hist.oneClass}
\title{hist.oneClass}
\usage{
\method{hist}{oneClass}(x, predUn = NULL, th = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class 'oneClass'.}

\item{predUn}{a vector of unlabeled predictions (if \code{NULL} \code{x$predUn} is used, if existing).}

\item{th}{draw vertical lines in the histogram, indication for a threshold.}

\item{ylim}{the y limits of the plot.}

\item{...}{other arguments that can be passed to \code{\link{plot}}.}
}
\value{
... (invisible).
}
\description{
... .
}
\details{
... .
}
\examples{
### to do
}

