% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_default_fc.R
\name{maxent_default_fc}
\alias{maxent_default_fc}
\title{maxent_default_fc}
\usage{
maxent_default_fc(nPos)
}
\arguments{
\item{nPos}{number of positive training samples}
}
\description{
...
}
\examples{
\dontrun{
maxent_default_fc(9)
maxent_default_fc(10)
maxent_default_fc(14)
maxent_default_fc(15)
maxent_default_fc(79)
maxent_default_fc(80)
}
}

