% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nmm_rank_models.R
\name{nmm_rank_models}
\alias{nmm_rank_models}
\title{nmm_rank_models}
\usage{
nmm_rank_models(x)
}
\arguments{
\item{x}{the table returned from \code{\link{nmm_selection_criteria}}}
}
\description{
Ranks the models (rows) according to different criteria (columns) and calculates the rank-sum. It is assumed that the input \code{x} has been derived from the function \code{\link{nmm_selection_criteria}}.
}

