% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nmm_rank_models_iteratively}
\alias{nmm_rank_models_iteratively}
\title{nmm_rank_models_iteratively}
\usage{
nmm_rank_models_iteratively(x)
}
\arguments{
\item{x}{the table returned from \code{\link{nmm_selection_criteria}}}
}
\description{
Iteratively ranks the models (rows) according to different criteria (columns)  with \code{\link{nmm_rank_models}}, removes the model with the lowest rank-sum. These two steps are repeated until the last model is ranked (at the first place). It is assumed that the input \code{x} has been derived from the function \code{\link{nmm_selection_criteria}}.
}

