% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.ModelEvaluation}
\alias{plot}
\alias{plot.ModelEvaluation}
\title{Plot PU-performance vs. evaluation accuracies to evaluate model selection.}
\usage{
\method{plot}{ModelEvaluation}(x, ...)
}
\arguments{
\item{x}{an object of class ModelSelectionEvaluation
(see \code{\link{evaluate}}).}

\item{...}{other arguments passed tot \code{\link{plot}}.}
}
\description{
...
}
\examples{
### to do
}

