% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.ModelEvaluation}
\alias{plot.ModelEvaluation}
\title{Plot accuracy statistics dependent on threshold.}
\usage{
\method{plot}{ModelEvaluation}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[dismo]{ModelEvaluation}}.
(see \code{\link{evaluate}}).}

\item{...}{other arguments passed to \code{\link{plot}}.}
}
\description{
Kappa/Sensitivity (Producer's Accuracy) and Positive Predictive Value (Users's Accuracy) are plotted.
}
\examples{
\dontrun{
# get a ModelEvaluation object
ev <- dismo::evaluate( p=rnorm(25, mean = 1), a=rnorm(100, mean = -1) )
plot(ev)
}
}

