% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PPPvsTPR.R
\name{plot_PPPvsTPR}
\alias{plot_PPPvsTPR}
\title{Plot of the probability of positive prediction versus the true positive rate.}
\usage{
plot_PPPvsTPR(x, identifyPoints = FALSE, add = FALSE,
  highlightBest = TRUE, xlim = c(0, 1), ylim = c(0, 1), ...)
}
\arguments{
\item{x}{a \code{trainOcc} object}

\item{identifyPoints}{logical with default \code{FALSE}. set to \code{TRUE} and click 
close to the points in the plot if you want to find out the model rows}

\item{add}{logical, add the plot to another plot}

\item{highlightBest}{logical with default set to \code{TRUE} which highlights the model ranked highest by the selection criteria, i.e. the one in \code{x$metric}}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{...}{other parameters passed to plot (if \code{add} is \code{FALSE}) or points (if \code{add} is \code{TRUE})}
}
\value{
a scatterplot. the point corresponding to the final model in \code{x} is highlighted.
}
\description{
Assumption that the true positive rate can be estimated well from the 
positive data, it is reasonable to assume that for a given true positive rate a model 
which leads to a lower probability of positive predictions is more accurate. This is because 
it will lead to lower fals positive predictions.
}
\details{
plot_PPPvsTPR
}
\examples{
\dontrun{
data(bananas)
### get some models: 
# one-class svm
ocsvm <- trainOcc (x = bananas$tr[, -1], y = bananas$tr[, 1], method="ocsvm", 
                   tuneGrid=expand.grid(sigma=c(seq(0.01, .09, .02), seq(0.1, .9, .1)), 
                                        nu=seq(.05,.55,.1)) )
# biased svm
biasedsvm <- trainOcc (x = bananas$tr[, -1], y = bananas$tr[, 1], method="bsvm", 
                       tuneGrid=expand.grid(sigma=c(0.1, 1), 
                                            cNeg=2^seq(-4, 2, 2), 
                                            cMultiplier=2^seq(4, 8, 2) ) )
# compare with PPPvsTPR-plot
plot_PPPvsTPR(ocsvm)
plot_PPPvsTPR(biasedsvm, add=TRUE, col="red")
}
}

