% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ModelSelectionEvaluation.R
\name{print.ModelSelectionEvaluation}
\alias{print.ModelSelectionEvaluation}
\title{Print an object of class \code{ModelSelectionEvaluation}}
\usage{
\method{print}{ModelSelectionEvaluation}(x, by = NULL, digits = 2,
  invisible = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{ModelSelectionEvaluation}}

\item{by}{character giving a name of the table used to sort the table, default \code{NULL} is no sorting}

\item{digits}{number of digits to be printed}

\item{invisible}{do not print}

\item{...}{options passed to the generic print method.}
}
\description{
A table is printed with test and PU- performance metrics for the different evaluated models.
Test metrics: \cr
auc: auc
mxK.K: kappa at the threshold which leads to the maximum kappa
mxK.OA: Overall Accuracy at the threshold which leads to the maximum kappa
...
}

