% Generated by roxygen2 (4.0.1): do not edit by hand
\name{puSummary}
\alias{puSummary}
\title{puSummary}
\usage{
puSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{...}

\item{lev}{...}

\item{model}{...}
}
\value{
...
}
\description{
... .
}
\details{
... .
}
\examples{
\dontrun{
### create a PU-data example of random numbers
P <- rnorm( 30, 2 )
U <- c( rnorm( 80 ), rnorm( 20, 2 ) )
d <- data.frame( obs = puFactor( rep( c( 1, 0 ), c( 30, 100 ) ), positive = 1 ),
                 pos = c( P, U ) )
puSummary(d)
}
}

