% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamples.trainOcc.R
\name{resamples}
\alias{resamples}
\alias{resamples.trainOcc}
\title{Collation and Visualization of Resampling Results (Models of one Method!).}
\usage{
\method{resamples}{trainOcc}(x, modParam = NULL, modRow = NULL,
  modRank = NULL, metric = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{trainOcc}}.}

\item{modParam}{a data frame with the desired modParameters.}

\item{modRow}{the index of the model, i.e. the row in the  \code{object$results} table.}

\item{modRank}{the model at the modelRank-th position after sorting after the performance metric. The performance metric can be specified by \code{by}.}

\item{metric}{if modRank is used, which metric should be used for the ranking?}

\item{...}{...}
}
\description{
The function returns an object of class \code{\link{resamples}} as 
does the function \code{resamples.default} (see \code{\link[caret]{resamples}}). 
\code{resamples.default} can be used to collect, analyze, and visualize a resampling 
results of the final models from a list of \code{train} objects. 
Instead, \code{resamples.trainOcc} collects resampling results of different models 
from the same method stored in one \code{trainOcc} object. This is useful due to the fact that the 
metrics used for model selection are derived from positive and unlabeled data and always have to be 
treated with caution.
}
\examples{
\dontrun{
data(bananas)
oc <- trainOcc(x=bananas$tr[, -1], y=bananas$tr[, 1], 
               tuneGrid=expand.grid(sigma=c(0.1,1), ### not so large grid
                                    cNeg=2^seq(-5, 10, 3), 
                                    cMultiplier=2^seq(4, 15, 2)))
# visualize resamples of the ten models with highest (mean) puAuc values
resamps <- resamples(oc, modRank=1:10, metric="puAuc")
bwplot (resamps, metric='puF')
}
}

