% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_threeGaussians.R
\name{threeGaussians}
\alias{threeGaussians}
\title{Two-dimensional toy data consisting of four Gaussian 
distributions.}
\usage{
threeGaussians(n = 200, d = 2:4, n.tr = c(20, 100), seed = NULL)
}
\arguments{
\item{n}{integer. the number of test samples}

\item{d}{vector of three integers. the distance between the three negative gaussians and the positive gaussian}

\item{n.tr}{vector of two integer, i.e. the number of positive and unlabeled training samples}

\item{seed}{a seed point controlling the randomness.}
}
\value{
a list of the toy data set with two data frames (\code{tr} and 
\code{te}) both containing the columns \code{y}, \code{x1}, and \code{x2}.
in the training set the column \code{y} in \{0, 1\} indicates if a sample 
belongs to the positive class (1) or if it is unlabeled (0), and in the test 
set \code{te} if it belongs to the positive class (1) or the negative class 
(0). \code{x1} and \code{x2} are the features or predictors.
}
\examples{
\dontrun{
toy <- threeGaussians(seed=9)

### plot the test set
par(mfrow=c(1,2))
plot( toy$te[ , -1 ], pch = c( 4, 16 )[ toy$te$y + 1 ] )
legend('topright', legend=c('pos', 'neg'), pch=c(16, 4))

### plot the training set
plot( toy$tr[ , -1 ], pch = c( 4, 16 )[ toy$tr$y + 1 ] )
legend('topright', legend=c('pos', 'un'), pch=c(16, 4))
}
}

